Move new classes to separate file.
Additionally, a constants file was introduced,
in order to further clean up and centralize
their definitions.

A new property for gas prize strategy was added
to the Scenario class, as a drive-by optimization.
